The activation tests were causing the clean tests to fail.
This was because the activation tests need to override run to avoid actually activating the login shell.
However, my method of doing this was overriding Base.run which of course was messing up other parts of the code.
The correct solution was to use Mocking.jl to override that methods in a do block which will also force recompiling of the src
functions that use it.

Unfortunately, this means that tests for now will require manually installing Mocking.jl from github until it gets added to METADATA.