BuildExecutable.jl now handles ARGS.
Updated build.jl to make playground a binary executable and handles installation
of that executable a little bit more safely. Unfortunately, since there is no
`Pkg.install` or `Pkg.build("Playground", install=true)` we just check for a
PLAYGROUND_INSTALL variable that tells the build script to install the executable
and config to ~/.playground/bin/playground.